Add cached `dealloc` selector and optimize NSObject class fetching
