+++ crates/test-assembly/crates/test_declare_class/Cargo.toml
@@ -22,4 +22,4 @@ gnustep-2-0 = ["gnustep-1-9", "objc2?/gnustep-2-0"]
 gnustep-2-1 = ["gnustep-2-0", "objc2?/gnustep-2-1"]
 
 # Hack to prevent the feature flag from being enabled in the entire project
-assembly-features = ["objc2?/unstable-static-sel-inlined", "objc2?/unstable-static-class"]
+assembly-features = ["objc2?/unstable-static-sel-inlined", "objc2?/unstable-static-class-inlined"]
